+++ packages/components/organisms/f-checkout/src/components/Checkout.vue
@@ -154,6 +154,16 @@ export default {
             required: true
         },
 
+        placeOrderUrl: {
+            type: String,
+            required: true
+        },
+
+        paymentPageUrlPrefix: {
+            type: String,
+            required: true
+        },
+
         updateCheckoutUrl: {
             type: String,
             required: true
@@ -181,6 +191,11 @@ export default {
             default: 1000
         },
 
+        placeOrderTimeout: {
+            type: Number,
+            default: 1000
+        },
+
         updateCheckoutTimeout: {
             type: Number,
             default: 1000
@@ -204,6 +219,11 @@ export default {
         getAddressTimeout: {
             type: Number,
             default: 1000
+        },
+
+        applicationName: {
+            type: String,
+            required: true
         },
 
         getGeoLocationUrl: {
@@ -260,6 +280,8 @@ export default {
             'serviceType',
             'time',
             'userNote',
+            'basket',
+            'orderId',
             'geolocation'
         ]),
 
@@ -317,6 +339,7 @@ export default {
             'setAuthToken',
             'updateCustomerDetails',
             'updateUserNote',
+            'placeOrder',
             'getGeoLocation'
         ]),
 
@@ -374,6 +397,8 @@ export default {
                     timeout: this.updateCheckoutTimeout
                 });
 
+                await this.submitOrder();
+
                 this.$emit(EventNames.CheckoutSuccess, eventData);
 
                 this.$logger.logInfo(
@@ -381,6 +406,8 @@ export default {
                     this.$store,
                     eventData
                 );
+
+                this.redirectToPayment();
             } catch (thrownErrors) {
                 eventData.errors = thrownErrors;
 
@@ -392,6 +419,39 @@ export default {
                     eventData
                 );
             }
+        },
+
+        /**
+         * Redirect to the payment page.
+         */
+        redirectToPayment () {
+            setTimeout(() => { // TODO: remove this when the order team handles this automatically.
+                window.location.assign(`${this.paymentPageUrlPrefix}/${this.orderId}`);
+            }, 1000);
+        },
+
+        /**
+         * Place the order.
+         */
+        async submitOrder () {
+            const data = {
+                basketId: this.basket.id,
+                applicationId: 7, // Responsive Web
+                customerNotes: {
+                    noteForRestaurant: this.userNote
+                },
+                applicationName: this.applicationName,
+                applicationVersion: '1',
+                referralState: 'None',
+                deviceId: '127.0.0.1', // TODO: TBC
+                deviceName: window.navigator.userAgent
+            };
+
+            await this.placeOrder({
+                url: this.placeOrderUrl,
+                data,
+                timeout: this.placeOrderTimeout
+            });
         },
 
         /**
@@ -525,7 +585,11 @@ export default {
          *
          */
         async lookupGeoLocation () {
-            const locationData = { addressLines: Object.values(this.address).filter(addressLine => !!addressLine) };
+            const locationData =
+            {
+                addressLines: Object.values(this.address).filter(addressLine => !!addressLine)
+            };
+
             try {
                 if (locationData.addressLines.length) {
                     await this.getGeoLocation({