+++ packages/components/organisms/f-checkout/src/services/mapper.js
@@ -5,37 +5,49 @@ const mapUpdateCheckoutRequest = ({
     time,
     userNote,
     geolocation
-}) => ({
-    customer: {
-        firstName: customer.firstName || null,
-        lastName: customer.lastName || null,
-        phoneNumber: customer.mobileNumber || null,
-        dateOfBirth: null
-    },
-    fulfilment: {
-        time,
-        location: {
-            ...(isCheckoutMethodDelivery ? {
-                address: {
-                    lines: [
-                        address.line1 || '',
-                        address.line2 || '',
-                        '',
-                        address.city || '',
-                        ''
-                    ],
-                    postalCode: address.postcode || null
-                }
-            } : {}),
-            geolocation
+}) => ([
+    {
+        op: 'add',
+        path: '/customer',
+        value: {
+            firstName: customer.firstName || null,
+            lastName: customer.lastName || null,
+            phoneNumber: customer.mobileNumber || null,
+            dateOfBirth: null
         }
     },
-    notes: [
-        {
-            type: 'delivery',
-            note: userNote
+    {
+        op: 'add',
+        path: '/fulfilment',
+        value: {
+            time,
+            location: {
+                ...(isCheckoutMethodDelivery ? {
+                    address: {
+                        lines: [
+                            address.line1 || '',
+                            address.line2 || '',
+                            '',
+                            address.city || '',
+                            ''
+                        ],
+                        postalCode: address.postcode || null
+                    }
+                } : {}),
+                geolocation
+            }
         }
-    ]
-});
+    },
+    {
+        op: 'add',
+        path: '/notes',
+        value: [
+            {
+                type: 'delivery',
+                note: userNote
+            }
+        ]
+    }
+]);
 
 export default mapUpdateCheckoutRequest;