+++ packages/components/organisms/f-content-cards/src/components/cardTemplates/index.js
@@ -2,6 +2,7 @@ import FirstTimeCustomerCard from './FirstTimeCustomerCard.vue';
 import PromotionCard from './PromotionCard.vue';
 import PostOrderCard from './PostOrderCard.vue';
 import SkeletonLoader from './SkeletonLoader.vue';
+import StampCard1 from './StampCard1.vue';
 import TermsAndConditionsCard from './TermsAndConditionsCard.vue';
 import VoucherCard from './voucherCard/VoucherCard.vue';
 import GroupHeaderCard from './GroupHeaderCard.vue';
@@ -12,6 +13,7 @@ export {
     PromotionCard,
     PostOrderCard,
     SkeletonLoader,
+    StampCard1,
     TermsAndConditionsCard,
     VoucherCard,
     HomePromotionCard1,