+++ packages/components/organisms/f-footer/test/specs/accessibility/axe-accessibility.spec.js
@@ -1,15 +1,18 @@
-import { getAccessibilityTestResults } from '../../../../../../../test/utils/axe-helper';
-import FooterComponent from '../../../test-utils/component-objects/f-footer.component';
+const { getAccessibilityTestResults } = require('../../../../../../../test/utils/axe-helper');
+const Footer = require('../../../test-utils/component-objects/f-footer.component');
+const footer = new Footer();
 
 describe('Accessibility tests', () => {
-    beforeEach(() => {
-        browser.url('?path=/story/components-organisms--footer-component');
-        browser.switchToFrame(0);
-        FooterComponent.waitForFooter();
+
+    it('a11y - should test f-footer component WCAG compliance for AU Locale', () => {
+        // Act
+        footer.open('au');
+        const axeResults = getAccessibilityTestResults('f-footer');
     });
 
-    it('a11y - should test f-footer component WCAG compliance', () => {
+    it('a11y - should test f-footer component WCAG compliance for GB Locale', () => {
         // Act
+        footer.open();
         const axeResults = getAccessibilityTestResults('f-footer');
     });
 });